# openapi.yaml
openapi: 3.0.3
info:
  title: Gelos Grocery Orders REST API
  version: 1.0.0
  description: >
    REST API for managing order details for Gelos grocery store, secured with JWT.
servers:
  - url: http://localhost:3000
    description: Local development server

components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT

  schemas:
    Order:
      type: object
      required:
        - orderNumber
        - orderDate
        - customerNumber
        - productCode
        - productName
        - productQuantity
        - productPrice
        - totalAmount
        - modeOfPayment
      properties:
        _id:
          type: string
          description: MongoDB document ID
          readOnly: true
        orderNumber:
          type: integer
        orderDate:
          type: string
          example: "23/05/2022"
        customerNumber:
          type: integer
        productCode:
          type: integer
        productName:
          type: string
        productQuantity:
          type: integer
        productPrice:
          type: number
        totalAmount:
          type: number
        modeOfPayment:
          type: string
          example: "Cash"

    LoginRequest:
      type: object
      required:
        - username
        - password
      properties:
        username:
          type: string
          example: manager
        password:
          type: string
          example: manager123

    LoginResponse:
      type: object
      properties:
        token:
          type: string
          description: JWT access token

paths:
  /login:
    post:
      summary: Login and obtain JWT token
      tags:
        - Auth
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/LoginRequest'
      responses:
        '200':
          description: JWT token returned
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/LoginResponse'
        '401':
          description: Invalid credentials

  /orders:
    get:
      summary: Get all orders
      tags:
        - Orders
      security:
        - bearerAuth: []
      responses:
        '200':
          description: List of orders
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Order'
        '401':
          description: Missing or invalid token
    post:
      summary: Create a new order
      tags:
        - Orders
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Order'
      responses:
        '201':
          description: Order created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Order'
        '400':
          description: Validation error
        '401':
          description: Missing or invalid token
        '403':
          description: Insufficient permissions

  /orders/{id}:
    get:
      summary: Get a single order by ID
      tags:
        - Orders
      security:
        - bearerAuth: []
      parameters:
        - in: path
          name: id
          required: true
          description: Order ID
          schema:
            type: string
      responses:
        '200':
          description: Order found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Order'
        '401':
          description: Missing or invalid token
        '404':
          description: Order not found
    put:
      summary: Replace an existing order
      tags:
        - Orders
      security:
        - bearerAuth: []
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Order'
      responses:
        '200':
          description: Order updated
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Order'
        '400':
          description: Validation error
        '401':
          description: Missing or invalid token
        '403':
          description: Insufficient permissions
        '404':
          description: Order not found
    patch:
      summary: Partially update an order
      tags:
        - Orders
      security:
        - bearerAuth: []
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              description: Fields to update
      responses:
        '200':
          description: Order updated
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Order'
        '400':
          description: Validation error
        '401':
          description: Missing or invalid token
        '403':
          description: Insufficient permissions
        '404':
          description: Order not found
    delete:
      summary: Delete an order
      tags:
        - Orders
      security:
        - bearerAuth: []
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Order deleted successfully
        '401':
          description: Missing or invalid token
        '403':
          description: Insufficient permissions
        '404':
          description: Order not found